+++ DIRECTORY.md
@@ -1,17 +1,4 @@
 
-## Arithmetic Analysis
-  * [Bisection](arithmetic_analysis/bisection.py)
-  * [Gaussian Elimination](arithmetic_analysis/gaussian_elimination.py)
-  * [In Static Equilibrium](arithmetic_analysis/in_static_equilibrium.py)
-  * [Intersection](arithmetic_analysis/intersection.py)
-  * [Jacobi Iteration Method](arithmetic_analysis/jacobi_iteration_method.py)
-  * [Lu Decomposition](arithmetic_analysis/lu_decomposition.py)
-  * [Newton Forward Interpolation](arithmetic_analysis/newton_forward_interpolation.py)
-  * [Newton Method](arithmetic_analysis/newton_method.py)
-  * [Newton Raphson](arithmetic_analysis/newton_raphson.py)
-  * [Newton Raphson New](arithmetic_analysis/newton_raphson_new.py)
-  * [Secant Method](arithmetic_analysis/secant_method.py)
-
 ## Audio Filters
   * [Butterworth Filter](audio_filters/butterworth_filter.py)
   * [Iir Filter](audio_filters/iir_filter.py)
@@ -520,6 +507,9 @@
     * [Test Knapsack](knapsack/tests/test_knapsack.py)
 
 ## Linear Algebra
+  * [Gaussian Elimination](linear_algebra/gaussian_elimination.py)
+  * [Jacobi Iteration Method](linear_algebra/jacobi_iteration_method.py)
+  * [Lu Decomposition](linear_algebra/lu_decomposition.py)
   * Src
     * [Conjugate Gradient](linear_algebra/src/conjugate_gradient.py)
     * [Lib](linear_algebra/src/lib.py)
@@ -583,7 +573,6 @@
   * [Binary Multiplication](maths/binary_multiplication.py)
   * [Binomial Coefficient](maths/binomial_coefficient.py)
   * [Binomial Distribution](maths/binomial_distribution.py)
-  * [Bisection](maths/bisection.py)
   * [Ceil](maths/ceil.py)
   * [Chebyshev Distance](maths/chebyshev_distance.py)
   * [Check Polygon](maths/check_polygon.py)
@@ -617,7 +606,6 @@
   * [Germain Primes](maths/germain_primes.py)
   * [Greatest Common Divisor](maths/greatest_common_divisor.py)
   * [Hardy Ramanujanalgo](maths/hardy_ramanujanalgo.py)
-  * [Integration By Simpson Approx](maths/integration_by_simpson_approx.py)
   * [Interquartile Range](maths/interquartile_range.py)
   * [Is Int Palindrome](maths/is_int_palindrome.py)
   * [Is Ip V4 Address Valid](maths/is_ip_v4_address_valid.py)
@@ -644,10 +632,24 @@
   * [Modular Exponential](maths/modular_exponential.py)
   * [Monte Carlo](maths/monte_carlo.py)
   * [Monte Carlo Dice](maths/monte_carlo_dice.py)
-  * [Nevilles Method](maths/nevilles_method.py)
-  * [Newton Raphson](maths/newton_raphson.py)
   * [Number Of Digits](maths/number_of_digits.py)
-  * [Numerical Integration](maths/numerical_integration.py)
+  * Numerical Analysis
+    * [Bisection](maths/numerical_analysis/bisection.py)
+    * [Bisection 2](maths/numerical_analysis/bisection_2.py)
+    * [Integration By Simpson Approx](maths/numerical_analysis/integration_by_simpson_approx.py)
+    * [Intersection](maths/numerical_analysis/intersection.py)
+    * [Nevilles Method](maths/numerical_analysis/nevilles_method.py)
+    * [Newton Forward Interpolation](maths/numerical_analysis/newton_forward_interpolation.py)
+    * [Newton Method](maths/numerical_analysis/newton_method.py)
+    * [Newton Raphson](maths/numerical_analysis/newton_raphson.py)
+    * [Newton Raphson 2](maths/numerical_analysis/newton_raphson_2.py)
+    * [Newton Raphson New](maths/numerical_analysis/newton_raphson_new.py)
+    * [Numerical Integration](maths/numerical_analysis/numerical_integration.py)
+    * [Runge Kutta](maths/numerical_analysis/runge_kutta.py)
+    * [Runge Kutta Fehlberg 45](maths/numerical_analysis/runge_kutta_fehlberg_45.py)
+    * [Secant Method](maths/numerical_analysis/secant_method.py)
+    * [Simpson Rule](maths/numerical_analysis/simpson_rule.py)
+    * [Square Root](maths/numerical_analysis/square_root.py)
   * [Odd Sieve](maths/odd_sieve.py)
   * [Perfect Cube](maths/perfect_cube.py)
   * [Perfect Number](maths/perfect_number.py)
@@ -673,8 +675,6 @@
   * [Radians](maths/radians.py)
   * [Radix2 Fft](maths/radix2_fft.py)
   * [Remove Digit](maths/remove_digit.py)
-  * [Runge Kutta](maths/runge_kutta.py)
-  * [Runge Kutta Fehlberg 45](maths/runge_kutta_fehlberg_45.py)
   * [Segmented Sieve](maths/segmented_sieve.py)
   * Series
     * [Arithmetic](maths/series/arithmetic.py)
@@ -687,7 +687,6 @@
   * [Sieve Of Eratosthenes](maths/sieve_of_eratosthenes.py)
   * [Sigmoid](maths/sigmoid.py)
   * [Signum](maths/signum.py)
-  * [Simpson Rule](maths/simpson_rule.py)
   * [Simultaneous Linear Equation Solver](maths/simultaneous_linear_equation_solver.py)
   * [Sin](maths/sin.py)
   * [Sock Merchant](maths/sock_merchant.py)
@@ -709,7 +708,6 @@
     * [Proth Number](maths/special_numbers/proth_number.py)
     * [Ugly Numbers](maths/special_numbers/ugly_numbers.py)
     * [Weird Number](maths/special_numbers/weird_number.py)
-  * [Square Root](maths/square_root.py)
   * [Sum Of Arithmetic Series](maths/sum_of_arithmetic_series.py)
   * [Sum Of Digits](maths/sum_of_digits.py)
   * [Sum Of Geometric Progression](maths/sum_of_geometric_progression.py)
@@ -812,6 +810,7 @@
   * [Horizontal Projectile Motion](physics/horizontal_projectile_motion.py)
   * [Hubble Parameter](physics/hubble_parameter.py)
   * [Ideal Gas Law](physics/ideal_gas_law.py)
+  * [In Static Equilibrium](physics/in_static_equilibrium.py)
   * [Kinetic Energy](physics/kinetic_energy.py)
   * [Lorentz Transformation Four Vector](physics/lorentz_transformation_four_vector.py)
   * [Malus Law](physics/malus_law.py)
