+++ maths/numerical_analysis/bisection.py
