+++ maths/numerical_analysis/bisection_2.py
