+++ maths/numerical_analysis/integration_by_simpson_approx.py
