+++ maths/numerical_analysis/intersection.py
