+++ maths/numerical_analysis/newton_forward_interpolation.py
