+++ maths/numerical_analysis/newton_method.py
