+++ maths/numerical_analysis/secant_method.py
