+++ maths/numerical_analysis/simpson_rule.py
